Move DB to services
